Add a sudo layer (Bug #2243)
SGE.Actions.Disable_Or_Enable gets a Boolean parameter that selects whether sudo should be called. In principle, this could be deduced from the value of Env(REMOTE_USER): if it is empty, do not use sudo. However, setting this to True explicitly is a safety feature, so calls that are never meant to be executed by a web user will not be issued even if REMOTE_USER is somehow set erroneously.
Of course, a web user may well be an admin, so we might want to execute (almost) all possible calls from the CGI anyway. If that turns out to be the case, removing the Boolean and automating the sudo layer is still possible.