Fix Posix Priority (Bug #1341)
Posix priority is given different values in qstat -j and qstat -u (-xml): -1023..1024, the official scale, in one case, but shifted +1024 in the other.
This means reduced re-usability of code across different qstat invocations, so introduce a (somewhat ugly) Boolean parameter so the inner routines know whether priority has to be shifted.